refactor(server): split PR feedback selectors
