+++ crates/harness-server/src/workflow_runtime_pr_feedback.rs
@@ -16,7 +16,11 @@ use std::path::Path;
 
 const DEFAULT_PR_FEEDBACK_FAILED_CHILD_SUPPRESSION_SECS: u64 = 24 * 60 * 60;
 
+mod command_state;
 mod pr_lifecycle_persist;
+mod targets;
+
+use command_state::*;
 use pr_lifecycle_persist::{
     issue_workflow_id, persist_pr_lifecycle_with_retry, pr_lifecycle_workflow_id,
 };
@@ -24,6 +28,7 @@ use pr_lifecycle_persist::{
 use pr_lifecycle_persist::{
     set_pr_lifecycle_persist_test_failures, PR_LIFECYCLE_PERSIST_MAX_ATTEMPTS,
 };
+use targets::*;
 
 pub(crate) struct PrDetectedRuntimeContext<'a> {
     pub project_root: &'a Path,
@@ -1269,408 +1274,6 @@ async fn commit_runtime_decision(
         Some(_) => RuntimeDecisionCommitOutcome::Accepted,
         None => RuntimeDecisionCommitOutcome::Stale,
     })
-}
-
-fn is_active_pr_feedback_command_status(status: WorkflowCommandStatus) -> bool {
-    matches!(
-        status,
-        WorkflowCommandStatus::Pending
-            | WorkflowCommandStatus::Dispatching
-            | WorkflowCommandStatus::Deferred
-            | WorkflowCommandStatus::Dispatched
-    )
-}
-
-async fn has_active_local_review_command(
-    store: &WorkflowRuntimeStore,
-    workflow_id: &str,
-) -> anyhow::Result<bool> {
-    Ok(store
-        .commands_for(workflow_id)
-        .await?
-        .into_iter()
-        .any(|record| {
-            is_active_pr_feedback_command_status(record.status)
-                && record.command.activity_name() == Some(LOCAL_REVIEW_ACTIVITY)
-        }))
-}
-
-#[cfg(test)]
-async fn has_active_pr_feedback_command(
-    store: &WorkflowRuntimeStore,
-    workflow_id: &str,
-    failed_child_suppression_secs: u64,
-) -> anyhow::Result<bool> {
-    has_active_pr_feedback_command_with_activity(
-        store,
-        workflow_id,
-        failed_child_suppression_secs,
-        None,
-    )
-    .await
-}
-
-async fn has_active_pr_feedback_command_with_activity(
-    store: &WorkflowRuntimeStore,
-    workflow_id: &str,
-    failed_child_suppression_secs: u64,
-    latest_pr_activity_at: Option<chrono::DateTime<chrono::Utc>>,
-) -> anyhow::Result<bool> {
-    let parent_has_active_command =
-        store
-            .commands_for(workflow_id)
-            .await?
-            .into_iter()
-            .any(|record| {
-                is_active_pr_feedback_command_status(record.status)
-                    && matches!(
-                        record.command.activity_name(),
-                        Some("sweep_pr_feedback" | "address_pr_feedback")
-                    )
-                    || is_active_pr_feedback_command_status(record.status)
-                        && record.command.command_type
-                            == harness_workflow::runtime::WorkflowCommandType::StartChildWorkflow
-                        && record
-                            .command
-                            .command
-                            .get("definition_id")
-                            .and_then(|value| value.as_str())
-                            == Some(PR_FEEDBACK_DEFINITION_ID)
-            });
-    if parent_has_active_command {
-        return Ok(true);
-    }
-
-    // Scope the child lookup to this parent's children via the indexed-by-value
-    // `parent_workflow_id` column, rather than loading every PR-feedback instance
-    // across all projects and filtering in memory (which scales with the whole
-    // table and inflates memory use).
-    for instance in store
-        .list_instances_by_parent(workflow_id, None)
-        .await?
-        .into_iter()
-        .filter(|instance| instance.definition_id == PR_FEEDBACK_DEFINITION_ID)
-    {
-        if matches!(instance.state.as_str(), "pending" | "inspecting")
-            && has_active_child_pr_feedback_command(store, &instance.id).await?
-        {
-            return Ok(true);
-        }
-        if matches!(instance.state.as_str(), "failed" | "blocked")
-            && has_recent_failed_child_pr_feedback_command(
-                store,
-                &instance.id,
-                failed_child_suppression_secs,
-                latest_pr_activity_at,
-            )
-            .await?
-        {
-            return Ok(true);
-        }
-    }
-
-    Ok(false)
-}
-
-async fn has_recent_failed_child_pr_feedback_command(
-    store: &WorkflowRuntimeStore,
-    workflow_id: &str,
-    suppression_secs: u64,
-    latest_pr_activity_at: Option<chrono::DateTime<chrono::Utc>>,
-) -> anyhow::Result<bool> {
-    let Some(cutoff) = failed_child_suppression_cutoff(suppression_secs) else {
-        return Ok(false);
-    };
-    Ok(store
-        .commands_for(workflow_id)
-        .await?
-        .into_iter()
-        .any(|record| {
-            matches!(
-                record.status,
-                WorkflowCommandStatus::Failed | WorkflowCommandStatus::Blocked
-            ) && record.command.activity_name() == Some(PR_FEEDBACK_INSPECT_ACTIVITY)
-                && record.updated_at >= cutoff
-                && failed_child_suppression_still_applies(record.updated_at, latest_pr_activity_at)
-        }))
-}
-
-fn failed_child_suppression_still_applies(
-    failed_command_updated_at: chrono::DateTime<chrono::Utc>,
-    latest_pr_activity_at: Option<chrono::DateTime<chrono::Utc>>,
-) -> bool {
-    latest_pr_activity_at
-        .map(|activity_at| activity_at < failed_command_updated_at)
-        .unwrap_or(true)
-}
-
-fn failed_child_suppression_cutoff(suppression_secs: u64) -> Option<chrono::DateTime<chrono::Utc>> {
-    if suppression_secs == 0 {
-        return None;
-    }
-    let now = chrono::Utc::now();
-    Some(
-        i64::try_from(suppression_secs)
-            .ok()
-            .and_then(chrono::Duration::try_seconds)
-            .and_then(|duration| now.checked_sub_signed(duration))
-            .unwrap_or(chrono::DateTime::<chrono::Utc>::MIN_UTC),
-    )
-}
-
-async fn has_active_child_pr_feedback_command(
-    store: &WorkflowRuntimeStore,
-    workflow_id: &str,
-) -> anyhow::Result<bool> {
-    Ok(store
-        .commands_for(workflow_id)
-        .await?
-        .into_iter()
-        .any(|record| {
-            is_active_pr_feedback_command_status(record.status)
-                && record.command.activity_name() == Some(PR_FEEDBACK_INSPECT_ACTIVITY)
-        }))
-}
-
-async fn upsert_github_issue_pr_definition(store: &WorkflowRuntimeStore) -> anyhow::Result<()> {
-    store
-        .upsert_definition(&WorkflowDefinition::new(
-            GITHUB_ISSUE_PR_DEFINITION_ID,
-            1,
-            "GitHub issue PR workflow",
-        ))
-        .await
-}
-
-async fn load_or_issue_instance(
-    store: &WorkflowRuntimeStore,
-    workflow_id: String,
-    project_id: String,
-    repo: Option<String>,
-    issue_number: u64,
-    state: &str,
-) -> anyhow::Result<(WorkflowInstance, bool)> {
-    Ok(match store.get_instance(&workflow_id).await? {
-        Some(instance) => (instance, false),
-        None => (
-            issue_instance(workflow_id, project_id, repo, issue_number, state),
-            true,
-        ),
-    })
-}
-
-async fn load_or_pr_runtime_target(
-    store: &WorkflowRuntimeStore,
-    project_root: &Path,
-    repo: Option<&str>,
-    issue_number: Option<u64>,
-    pr_number: u64,
-    task_id: &TaskId,
-    pr_url: Option<&str>,
-    state: &str,
-) -> anyhow::Result<PrRuntimeTarget> {
-    upsert_github_issue_pr_definition(store).await?;
-    let project_id = project_root.to_string_lossy().into_owned();
-    if let Some(issue_number) = issue_number {
-        let workflow_id =
-            harness_workflow::issue_lifecycle::workflow_id(&project_id, repo, issue_number);
-        let (instance, new_instance) = load_or_issue_instance(
-            store,
-            workflow_id,
-            project_id,
-            repo.map(ToOwned::to_owned),
-            issue_number,
-            state,
-        )
-        .await?;
-        return Ok(PrRuntimeTarget {
-            instance,
-            new_instance,
-            issue_number: Some(issue_number),
-        });
-    }
-
-    if let Some(instance) = store
-        .get_instance_by_pr(GITHUB_ISSUE_PR_DEFINITION_ID, &project_id, repo, pr_number)
-        .await?
-    {
-        let issue_number = instance
-            .data
-            .get("issue_number")
-            .and_then(serde_json::Value::as_u64);
-        return Ok(PrRuntimeTarget {
-            instance,
-            new_instance: false,
-            issue_number,
-        });
-    }
-
-    Ok(PrRuntimeTarget {
-        instance: pr_scoped_instance(
-            pr_workflow_id(&project_id, repo, pr_number),
-            project_id,
-            repo.map(ToOwned::to_owned),
-            task_id,
-            pr_number,
-            pr_url,
-            state,
-        ),
-        new_instance: true,
-        issue_number: None,
-    })
-}
-
-fn issue_instance(
-    workflow_id: String,
-    project_id: String,
-    repo: Option<String>,
-    issue_number: u64,
-    state: &str,
-) -> WorkflowInstance {
-    WorkflowInstance::new(
-        "github_issue_pr",
-        1,
-        state,
-        WorkflowSubject::new("issue", format!("issue:{issue_number}")),
-    )
-    .with_id(workflow_id)
-    .with_data(crate::workflow_runtime_policy::merge_runtime_retry_policy(
-        Path::new(&project_id),
-        json!({
-            "project_id": project_id,
-            "repo": repo,
-            "issue_number": issue_number,
-        }),
-    ))
-}
-
-fn pr_scoped_instance(
-    workflow_id: String,
-    project_id: String,
-    repo: Option<String>,
-    task_id: &TaskId,
-    pr_number: u64,
-    pr_url: Option<&str>,
-    state: &str,
-) -> WorkflowInstance {
-    let data = pr_runtime_data(
-        Path::new(&project_id),
-        project_id.clone(),
-        repo.as_deref(),
-        None,
-        task_id,
-        pr_number,
-        pr_url,
-        None,
-    );
-    WorkflowInstance::new(
-        GITHUB_ISSUE_PR_DEFINITION_ID,
-        1,
-        state,
-        WorkflowSubject::new("pr", format!("pr:{pr_number}")),
-    )
-    .with_id(workflow_id)
-    .with_data(data)
-}
-
-fn pr_workflow_id(project_id: &str, repo: Option<&str>, pr_number: u64) -> String {
-    format!(
-        "{project_id}::repo:{}::pr:{pr_number}:feedback",
-        repo.unwrap_or("<none>")
-    )
-}
-
-fn pr_runtime_data(
-    project_root: &Path,
-    project_id: String,
-    repo: Option<&str>,
-    issue_number: Option<u64>,
-    task_id: &TaskId,
-    pr_number: u64,
-    pr_url: Option<&str>,
-    feedback_summary: Option<&str>,
-) -> serde_json::Value {
-    let mut data = json!({
-        "project_id": project_id,
-        "repo": repo,
-        "task_id": task_id.as_str(),
-        "pr_number": pr_number,
-        "pr_url": pr_url,
-    });
-    if let Some(issue_number) = issue_number {
-        data["issue_number"] = json!(issue_number);
-    }
-    if let Some(feedback_summary) = feedback_summary {
-        data["feedback_summary"] = json!(feedback_summary);
-    }
-    crate::workflow_runtime_policy::merge_runtime_retry_policy(project_root, data)
-}
-
-fn merge_last_decision(mut data: serde_json::Value, decision: &str) -> serde_json::Value {
-    if let Some(object) = data.as_object_mut() {
-        object.insert("last_decision".to_string(), json!(decision));
-    }
-    data
-}
-
-fn merge_runtime_merge_data(
-    mut data: serde_json::Value,
-    decision: &str,
-    task_id: Option<&str>,
-    delete_branch: bool,
-    require_review_threads_resolved: bool,
-    require_clean_merge_state: bool,
-) -> serde_json::Value {
-    if let Some(object) = data.as_object_mut() {
-        object.insert("last_decision".to_string(), json!(decision));
-        object.insert("merge_approved_at".to_string(), json!(chrono::Utc::now()));
-        object.insert("merge_delete_branch".to_string(), json!(delete_branch));
-        object.insert(
-            "merge_require_review_threads_resolved".to_string(),
-            json!(require_review_threads_resolved),
-        );
-        object.insert(
-            "merge_require_clean_merge_state".to_string(),
-            json!(require_clean_merge_state),
-        );
-        if let Some(task_id) = task_id {
-            object.insert("merge_approved_task_id".to_string(), json!(task_id));
-        }
-    }
-    data
-}
-
-fn optional_string_field(data: &serde_json::Value, field: &str) -> Option<String> {
-    data.get(field)
-        .and_then(|value| value.as_str())
-        .filter(|value| !value.trim().is_empty())
-        .map(ToOwned::to_owned)
-}
-
-fn optional_bool_field(data: &serde_json::Value, field: &str) -> Option<bool> {
-    data.get(field).and_then(|value| value.as_bool())
-}
-
-fn required_u64_field(data: &serde_json::Value, field: &str) -> anyhow::Result<u64> {
-    data.get(field)
-        .and_then(|value| value.as_u64())
-        .ok_or_else(|| anyhow::anyhow!("runtime issue workflow is missing {field}"))
-}
-
-fn event_type(outcome: PrFeedbackOutcome) -> &'static str {
-    match outcome {
-        PrFeedbackOutcome::BlockingFeedback => "FeedbackFound",
-        PrFeedbackOutcome::NoActionableFeedback => "NoFeedbackFound",
-        PrFeedbackOutcome::ReadyToMerge => "PrReadyToMerge",
-    }
-}
-
-fn outcome_label(outcome: PrFeedbackOutcome) -> &'static str {
-    match outcome {
-        PrFeedbackOutcome::BlockingFeedback => "blocking_feedback",
-        PrFeedbackOutcome::NoActionableFeedback => "no_actionable_feedback",
-        PrFeedbackOutcome::ReadyToMerge => "ready_to_merge",
-    }
 }
 
 #[cfg(test)]
